Refactor Autocomplete SuggestionHandler logic
