+++ img-txt_viewer.py
@@ -29,7 +29,6 @@ import shutil
 import ctypes
 import zipfile
 import subprocess
-from functools import partial
 
 
 # Standard Library - GUI
@@ -63,7 +62,7 @@ from main.scripts import (
     edit_panel,
     CropUI,
 )
-from main.scripts.Autocomplete import Autocomplete
+from main.scripts.Autocomplete import SuggestionHandler
 from main.scripts.PopUpZoom import PopUpZoom as PopUpZoom
 from main.scripts.OnnxTagger import OnnxTagger as OnnxTagger
 
@@ -104,7 +103,7 @@ class ImgTxtViewer:
         self.find_dupe_file = find_dupe_file.FindDupeFile()
         self.crop_ui = CropUI.CropInterface()
         self.onnx_tagger = OnnxTagger(self)
-        self.autocomplete = Autocomplete
+        self.autocomplete = SuggestionHandler(self)
 
         # Setup UI state
         self.ui_state = "ImgTxtViewer"
@@ -255,8 +254,6 @@ class ImgTxtViewer:
         self.suggestion_quantity_var = IntVar(value=4)
         self.suggestion_threshold_var = StringVar(value="Normal")
         self.last_word_match_var = BooleanVar(value=False)
-        self.selected_suggestion_index = 0
-        self.suggestions = []
 
 
 # --------------------------------------
@@ -315,7 +312,7 @@ class ImgTxtViewer:
         self.optionsMenu.add_cascade(label="Options", underline=0, state="disable", menu=self.options_subMenu)
         self.options_subMenu.add_checkbutton(label="Clean-Text", underline=0, variable=self.cleaning_text_var, command=self.toggle_list_menu)
         self.options_subMenu.add_checkbutton(label="Auto-Delete Blank Files", underline=0, variable=self.auto_delete_blank_files_var)
-        self.options_subMenu.add_checkbutton(label="Colored Suggestions", underline=1, variable=self.colored_suggestion_var, command=self.update_autocomplete_dictionary)
+        self.options_subMenu.add_checkbutton(label="Colored Suggestions", underline=1, variable=self.colored_suggestion_var, command=self.autocomplete.update_autocomplete_dictionary)
         self.options_subMenu.add_checkbutton(label="Highlight Selection", underline=0, variable=self.highlight_selection_var)
         self.options_subMenu.add_checkbutton(label="Big Save Button", underline=0, variable=self.big_save_button_var, command=self.toggle_save_button_height)
         self.options_subMenu.add_checkbutton(label="List View", underline=0, variable=self.list_mode_var, command=self.toggle_list_mode)
@@ -362,25 +359,25 @@ class ImgTxtViewer:
         # Suggestion Dictionary Menu
         dictionaryMenu = Menu(autocompleteSettingsMenu, tearoff=0)
         autocompleteSettingsMenu.add_cascade(label="Dictionary", underline=11, menu=dictionaryMenu)
-        dictionaryMenu.add_checkbutton(label="English Dictionary", underline=0, variable=self.csv_english_dictionary, command=self.update_autocomplete_dictionary)
-        dictionaryMenu.add_checkbutton(label="Danbooru", underline=0, variable=self.csv_danbooru, command=self.update_autocomplete_dictionary)
-        dictionaryMenu.add_checkbutton(label="Danbooru (Safe)", underline=0, variable=self.csv_danbooru_safe, command=self.update_autocomplete_dictionary)
-        dictionaryMenu.add_checkbutton(label="Derpibooru", underline=0, variable=self.csv_derpibooru, command=self.update_autocomplete_dictionary)
-        dictionaryMenu.add_checkbutton(label="e621", underline=0, variable=self.csv_e621, command=self.update_autocomplete_dictionary)
+        dictionaryMenu.add_checkbutton(label="English Dictionary", underline=0, variable=self.csv_english_dictionary, command=self.autocomplete.update_autocomplete_dictionary)
+        dictionaryMenu.add_checkbutton(label="Danbooru", underline=0, variable=self.csv_danbooru, command=self.autocomplete.update_autocomplete_dictionary)
+        dictionaryMenu.add_checkbutton(label="Danbooru (Safe)", underline=0, variable=self.csv_danbooru_safe, command=self.autocomplete.update_autocomplete_dictionary)
+        dictionaryMenu.add_checkbutton(label="Derpibooru", underline=0, variable=self.csv_derpibooru, command=self.autocomplete.update_autocomplete_dictionary)
+        dictionaryMenu.add_checkbutton(label="e621", underline=0, variable=self.csv_e621, command=self.autocomplete.update_autocomplete_dictionary)
         dictionaryMenu.add_separator()
-        dictionaryMenu.add_command(label="Clear Selection", underline=0, command=self.clear_dictionary_csv_selection)
+        dictionaryMenu.add_command(label="Clear Selection", underline=0, command=self.autocomplete.clear_dictionary_csv_selection)
 
         # Suggestion Threshold Menu
         suggestion_threshold_menu = Menu(autocompleteSettingsMenu, tearoff=0)
         autocompleteSettingsMenu.add_cascade(label="Threshold", underline=11, menu=suggestion_threshold_menu)
         for level in ["Slow", "Normal", "Fast", "Faster"]:
-            suggestion_threshold_menu.add_radiobutton(label=level, variable=self.suggestion_threshold_var, value=level, command=self.set_suggestion_threshold)
+            suggestion_threshold_menu.add_radiobutton(label=level, variable=self.suggestion_threshold_var, value=level, command=self.autocomplete.set_suggestion_threshold)
 
         # Suggestion Quantity Menu
         suggestion_quantity_menu = Menu(autocompleteSettingsMenu, tearoff=0)
         autocompleteSettingsMenu.add_cascade(label="Quantity", underline=11, menu=suggestion_quantity_menu)
         for quantity in range(0, 10):
-            suggestion_quantity_menu.add_radiobutton(label=str(quantity), variable=self.suggestion_quantity_var, value=quantity, command=lambda suggestion_quantity=quantity: self.set_suggestion_quantity(suggestion_quantity))
+            suggestion_quantity_menu.add_radiobutton(label=str(quantity), variable=self.suggestion_quantity_var, value=quantity, command=lambda suggestion_quantity=quantity: self.autocomplete.set_suggestion_quantity(suggestion_quantity))
 
         # Match Mode Menu
         match_mode_menu = Menu(autocompleteSettingsMenu, tearoff=0)
@@ -1592,13 +1589,13 @@ class ImgTxtViewer:
         # Mouse binds
         self.text_box.bind("<Double-1>", lambda event: self.custom_select_word_for_text(event, self.text_box))
         self.text_box.bind("<Triple-1>", lambda event: self.custom_select_line_for_text(event, self.text_box))
-        self.text_box.bind("<Button-1>", lambda event: (self.remove_tag(), self.clear_suggestions()))
+        self.text_box.bind("<Button-1>", lambda event: (self.remove_tag(), self.autocomplete.clear_suggestions()))
         self.text_box.bind("<Button-2>", lambda event: (self.delete_tag_under_mouse(event), self.sync_title_with_content(event)))
         self.text_box.bind("<Button-3>", lambda event: (self.show_text_context_menu(event)))
         # Update the autocomplete suggestion label after every KeyRelease event.
-        self.text_box.bind("<KeyRelease>", lambda event: (self.update_suggestions(event), self.sync_title_with_content(event)))
+        self.text_box.bind("<KeyRelease>", lambda event: (self.autocomplete.update_suggestions(event), self.sync_title_with_content(event)))
         # Insert a newline after inserting an autocomplete suggestion when list_mode is active.
-        self.text_box.bind('<comma>', self.insert_newline_listmode)
+        self.text_box.bind('<comma>', self.autocomplete.insert_newline_listmode)
         # Highlight duplicates when selecting text with keyboard or mouse.
         self.text_box.bind("<Shift-Right>", lambda event: self.highlight_duplicates(event, mouse=False))
         self.text_box.bind("<Shift-Left>", lambda event: self.highlight_duplicates(event, mouse=False))
@@ -1719,24 +1716,24 @@ class ImgTxtViewer:
         # Selected Dictionary
         dictionary_menu = Menu(suggestion_context_menu, tearoff=0)
         suggestion_context_menu.add_cascade(label="Dictionary", menu=dictionary_menu)
-        dictionary_menu.add_checkbutton(label="English Dictionary", underline=0, variable=self.csv_english_dictionary, command=self.update_autocomplete_dictionary)
-        dictionary_menu.add_checkbutton(label="Danbooru", underline=0, variable=self.csv_danbooru, command=self.update_autocomplete_dictionary)
-        dictionary_menu.add_checkbutton(label="Danbooru (Safe)", underline=0, variable=self.csv_danbooru_safe, command=self.update_autocomplete_dictionary)
-        dictionary_menu.add_checkbutton(label="Derpibooru", underline=0, variable=self.csv_derpibooru, command=self.update_autocomplete_dictionary)
-        dictionary_menu.add_checkbutton(label="e621", underline=0, variable=self.csv_e621, command=self.update_autocomplete_dictionary)
+        dictionary_menu.add_checkbutton(label="English Dictionary", underline=0, variable=self.csv_english_dictionary, command=self.autocomplete.update_autocomplete_dictionary)
+        dictionary_menu.add_checkbutton(label="Danbooru", underline=0, variable=self.csv_danbooru, command=self.autocomplete.update_autocomplete_dictionary)
+        dictionary_menu.add_checkbutton(label="Danbooru (Safe)", underline=0, variable=self.csv_danbooru_safe, command=self.autocomplete.update_autocomplete_dictionary)
+        dictionary_menu.add_checkbutton(label="Derpibooru", underline=0, variable=self.csv_derpibooru, command=self.autocomplete.update_autocomplete_dictionary)
+        dictionary_menu.add_checkbutton(label="e621", underline=0, variable=self.csv_e621, command=self.autocomplete.update_autocomplete_dictionary)
         dictionary_menu.add_separator()
-        dictionary_menu.add_command(label="Clear Selection", underline=0, command=self.clear_dictionary_csv_selection)
+        dictionary_menu.add_command(label="Clear Selection", underline=0, command=self.autocomplete.clear_dictionary_csv_selection)
         # Suggestion Threshold
         suggestion_threshold_menu = Menu(suggestion_context_menu, tearoff=0)
         suggestion_context_menu.add_cascade(label="Threshold", menu=suggestion_threshold_menu)
         threshold_levels = ["Slow", "Normal", "Fast", "Faster"]
         for level in threshold_levels:
-            suggestion_threshold_menu.add_radiobutton(label=level, variable=self.suggestion_threshold_var, value=level, command=self.set_suggestion_threshold)
+            suggestion_threshold_menu.add_radiobutton(label=level, variable=self.suggestion_threshold_var, value=level, command=self.autocomplete.set_suggestion_threshold)
         # Suggestion Quantity
         suggestion_quantity_menu = Menu(suggestion_context_menu, tearoff=0)
         suggestion_context_menu.add_cascade(label="Quantity", menu=suggestion_quantity_menu)
         for quantity in range(0, 10):
-            suggestion_quantity_menu.add_radiobutton(label=str(quantity), variable=self.suggestion_quantity_var, value=quantity, command=lambda suggestion_quantity=quantity: self.set_suggestion_quantity(suggestion_quantity))
+            suggestion_quantity_menu.add_radiobutton(label=str(quantity), variable=self.suggestion_quantity_var, value=quantity, command=lambda suggestion_quantity=quantity: self.autocomplete.set_suggestion_quantity(suggestion_quantity))
         # Match Mode
         match_mode_menu = Menu(suggestion_context_menu, tearoff=0)
         suggestion_context_menu.add_cascade(label="Match Mode", menu=match_mode_menu)
@@ -2328,225 +2325,6 @@ class ImgTxtViewer:
 
 #endregion
 ################################################################################################################################################
-#region - Autocomplete
-
-
-# --------------------------------------
-# Display Suggestions
-# --------------------------------------
-    def handle_suggestion_event(self, event):
-        keysym = event.keysym
-        if keysym == "Tab":
-            if self.selected_suggestion_index < len(self.suggestions):
-                selected_suggestion = self.suggestions[self.selected_suggestion_index]
-                if isinstance(selected_suggestion, tuple):
-                    selected_suggestion = selected_suggestion[0]
-                self.insert_selected_suggestion(selected_suggestion.strip())
-            self.clear_suggestions()
-        elif keysym in ("Alt_L", "Alt_R"):
-            if self.suggestions and not self.is_alt_arrow_pressed:
-                self.selected_suggestion_index = (self.selected_suggestion_index - 1) % len(self.suggestions) if keysym == "Alt_R" else (self.selected_suggestion_index + 1) % len(self.suggestions)
-                self.highlight_suggestions()
-            self.is_alt_arrow_pressed = False
-        elif keysym in ("Up", "Down", "Left", "Right") or event.char == ",":
-            self.clear_suggestions()
-        else:
-            return False
-        return True
-
-
-    def update_suggestions(self, event=None):
-        if event is None:
-            event = type('', (), {})()
-            event.keysym = ''
-            event.char = ''
-        cursor_position = self.text_box.index("insert")
-        if self.cursor_inside_tag(cursor_position):
-            self.clear_suggestions()
-            return
-        if self.handle_suggestion_event(event):
-            return
-        text = self.text_box.get("1.0", "insert")
-        self.clear_suggestions()
-        if self.last_word_match_var.get():
-            words = text.split()
-            current_word = words[-1] if words else ''
-        else:
-            if self.list_mode_var.get():
-                elements = [element.strip() for element in text.split('\n')]
-            else:
-                elements = [element.strip() for element in text.split(',')]
-            current_word = elements[-1]
-        current_word = current_word.strip()
-        if current_word and (len(self.selected_csv_files) >= 1 or self.use_mytags_var.get()):
-            suggestions = self.autocomplete.get_suggestion(current_word)
-            suggestions.sort(key=lambda x: self.autocomplete.get_score(x[0], current_word), reverse=True)
-            self.suggestions = [(suggestion[0].replace("_", " ") if suggestion[0] not in self.autocomplete.tags_with_underscore else suggestion[0], suggestion[1]) for suggestion in suggestions]
-            if self.suggestions:
-                self.highlight_suggestions()
-            else:
-                self.clear_suggestions()
-        else:
-            self.clear_suggestions()
-
-
-    def highlight_suggestions(self):
-        def on_mouse_hover(tag_name, highlight, event=None):
-            if highlight:
-                widget.tag_config(tag_name, relief='raised', borderwidth=1)
-                widget.config(cursor="hand2")
-            else:
-                widget.tag_config(tag_name, relief='flat', borderwidth=0)
-                widget.config(cursor="")
-        widget = self.suggestion_textbox
-        widget.config(state='normal')
-        widget.delete('1.0', 'end')
-        configured_colors = set()
-        num_suggestions = len(self.suggestions)
-        for index, (suggestion_text, classifier_id) in enumerate(self.suggestions):
-            classifier_id = classifier_id[0]
-            color_index = int(classifier_id) % len(self.suggestion_colors) if classifier_id and classifier_id.isdigit() else 0
-            suggestion_color = self.suggestion_colors[color_index]
-            bullet_symbol = "⚫" if index == self.selected_suggestion_index else "⚪"
-            suggestion = f"suggestion_tag_{index}"
-            widget.insert('end', bullet_symbol)
-            widget.insert('end', f" {suggestion_text} ", (suggestion, suggestion_color))
-            if suggestion_color not in configured_colors:
-                widget.tag_config(suggestion_color, foreground=suggestion_color, font=('Segoe UI', '9'))
-                configured_colors.add(suggestion_color)
-            widget.tag_bind(suggestion, '<Button-1>', partial(self.on_suggestion_click, index))
-            widget.tag_bind(suggestion, '<Enter>', partial(on_mouse_hover, suggestion, True))
-            widget.tag_bind(suggestion, '<Leave>', partial(on_mouse_hover, suggestion, False))
-            if index < num_suggestions - 1:
-                widget.insert('end', ', ')
-        widget.config(state='disabled')
-
-
-    def cursor_inside_tag(self, cursor_position):
-        line, column = map(int, cursor_position.split('.'))
-        line_text = self.text_box.get(f"{line}.0", f"{line}.end")
-        if self.list_mode_var.get():
-            inside_tag = column not in (0, len(line_text))
-        else:
-            inside_tag = not (column == 0 or line_text[column-1:column] in (',', ' ') or line_text[column:column+1] in (',', ' ') or column == len(line_text))
-        return inside_tag
-
-
-    def clear_suggestions(self):
-        self.suggestions = []
-        self.selected_suggestion_index = 0
-        self.suggestion_textbox.config(state='normal')
-        self.suggestion_textbox.delete('1.0', 'end')
-        self.suggestion_textbox.insert('1.0', "...")
-        self.suggestion_textbox.config(state='disabled')
-
-
-# --------------------------------------
-# Insert Suggestion
-# --------------------------------------
-    def insert_selected_suggestion(self, selected_suggestion):
-        selected_suggestion = selected_suggestion.strip()
-        text = self.text_box.get("1.0", "insert").rstrip()
-        if self.last_word_match_var.get():
-            words = text.split()
-            current_word = words[-1] if words else ''
-        else:
-            elements = [element.strip() for element in text.split('\n' if self.list_mode_var.get() else ',')]
-            current_word = elements[-1]
-        remaining_text = self.text_box.get("insert", "end").rstrip('\n')
-        start_of_current_word = "1.0 + {} chars".format(len(text) - len(current_word))
-        self.text_box.delete(start_of_current_word, "insert")
-        if not remaining_text.startswith(('\n' if self.list_mode_var.get() else ',')) and not self.last_word_match_var.get():
-            self.text_box.insert(start_of_current_word, selected_suggestion + ('\n' if self.list_mode_var.get() else ','))
-        else:
-            self.text_box.insert(start_of_current_word, selected_suggestion)
-        if self.list_mode_var.get() and remaining_text:
-            self.insert_newline_listmode(called_from_insert=True)
-            self.text_box.mark_set("insert", "insert + 1 lines")
-
-
-    def position_cursor(self, start_of_current_word, selected_suggestion):
-        if self.text_box.get(start_of_current_word).startswith(' '):
-            offset = len(selected_suggestion) + 2
-        else:
-            offset = len(selected_suggestion) + 1
-        self.text_box.mark_set("insert", "{}+{}c".format(start_of_current_word, offset))
-        self.text_box.insert("insert", " ")
-
-
-    def insert_newline_listmode(self, event=None, called_from_insert=False):
-        if self.list_mode_var.get():
-            self.text_box.insert("insert", '\n')
-            if called_from_insert and self.text_box.index("insert") != self.text_box.index("end-1c"):
-                self.text_box.mark_set("insert", "insert-1l")
-            return 'break'
-
-
-    def on_suggestion_click(self, suggestion_index, event=None):
-        selected_suggestion, _ = self.suggestions[suggestion_index]
-        self.insert_selected_suggestion(selected_suggestion.strip())
-        self.clear_suggestions()
-
-
-# --------------------------------------
-# Suggestion Settings
-# --------------------------------------
-    def update_autocomplete_dictionary(self):
-        csv_vars = {
-            'danbooru.csv': self.csv_danbooru,
-            'danbooru_safe.csv': self.csv_danbooru_safe,
-            'e621.csv': self.csv_e621,
-            'dictionary.csv': self.csv_english_dictionary,
-            'derpibooru.csv': self.csv_derpibooru
-            }
-        self.selected_csv_files = [csv_file for csv_file, var in csv_vars.items() if var.get()]
-        if not self.selected_csv_files:
-            self.autocomplete = Autocomplete("None", include_my_tags=self.use_mytags_var.get())
-        else:
-            self.autocomplete = Autocomplete(self.selected_csv_files[0], include_my_tags=self.use_mytags_var.get())
-            for csv_file in self.selected_csv_files[1:]:
-                self.autocomplete.autocomplete_dict.update(Autocomplete(csv_file).autocomplete_dict, include_my_tags=self.use_mytags_var.get())
-        self.clear_suggestions()
-        self.set_suggestion_color(self.selected_csv_files[0] if self.selected_csv_files else "None")
-        self.set_suggestion_threshold()
-
-
-    def set_suggestion_color(self, csv_file):
-        color_mappings = {
-            'None':                 {0: "black"},
-            'dictionary.csv':       {0: "black",    1: "black",     2: "black",     3: "black",     4: "black",     5: "black",     6: "black",     7: "black",     8: "black"},
-            'danbooru.csv':         {0: "black",    1: "#c00004",   2: "black",     3: "#a800aa",   4: "#00ab2c",   5: "#fd9200"},
-            'danbooru_safe.csv':    {0: "black",    1: "#c00004",   2: "black",     3: "#a800aa",   4: "#00ab2c",   5: "#fd9200"},
-            'e621.csv':             {-1: "black",   0: "black",     1: "#f2ac08",   3: "#dd00dd",   4: "#00aa00",   5: "#ed5d1f",   6: "#ff3d3d",   7: "#ff3d3d",   8: "#228822"},
-            'derpibooru.csv':       {0: "black",    1: "#e5b021",   3: "#fd9961",   4: "#cf5bbe",   5: "#3c8ad9",   6: "#a6a6a6",   7: "#47abc1",   8: "#7871d0",   9: "#df3647",   10: "#c98f2b",  11: "#e87ebe"}
-        }
-        black_mappings = {key: "black" for key in color_mappings[csv_file].keys()}
-        self.suggestion_colors = color_mappings[csv_file] if self.colored_suggestion_var.get() else black_mappings
-
-
-    def set_suggestion_quantity(self, suggestion_quantity):
-        self.autocomplete.max_suggestions = suggestion_quantity
-        self.update_suggestions(event=None)
-
-
-    def set_suggestion_threshold(self):
-        thresholds = {
-            "Slow"  : 275000,
-            "Normal": 130000,
-            "Fast"  : 75000,
-            "Faster": 40000
-        }
-        self.autocomplete.suggestion_threshold = thresholds.get(self.suggestion_threshold_var.get())
-
-
-    def clear_dictionary_csv_selection(self):
-        for attr in ['csv_danbooru', 'csv_derpibooru', 'csv_e621', 'csv_english_dictionary']:
-            getattr(self, attr).set(False)
-        self.update_autocomplete_dictionary()
-
-
-#endregion
-################################################################################################################################################
 #region - TextBox Highlights
 
 
@@ -2831,7 +2609,7 @@ class ImgTxtViewer:
             self.primary_display_image.unbind("<Configure>")
             self.primary_display_image.bind("<Configure>", self.resize_and_scale_image_event)
             self.toggle_list_mode()
-            self.clear_suggestions()
+            self.autocomplete.clear_suggestions()
             self.highlight_custom_string()
             self.highlight_all_duplicates_var.set(False)
             self.update_thumbnail_panel()
@@ -3772,7 +3550,7 @@ class ImgTxtViewer:
             for tag in tags:
                 if tag.strip():
                     self.custom_dictionary_listbox.insert('end', tag.strip())
-            self.update_autocomplete_dictionary()
+            self.autocomplete.update_autocomplete_dictionary()
 
 
     def create_custom_dictionary(self, reset=False, refresh=True):
