+++ main/scripts/Autocomplete.py
@@ -12,11 +12,12 @@ import csv
 import pickle
 import itertools
 from collections import defaultdict
+from functools import partial
 
 
 #endregion
 ################################################################################################################################################
-#region CLS: Autocomplete
+#region CLS Autocomplete
 
 
 class Autocomplete:
@@ -217,3 +218,219 @@ class Autocomplete:
         if not classifier_id and suggestion.startswith(text[:3]):
             score += 1
         return score
+
+
+#endregion
+################################################################################################################################################
+#region CLS SuggestionHandler
+
+
+class SuggestionHandler:
+    def __init__(self, parent):
+        self.parent = parent
+        self.autocomplete = Autocomplete
+        self.suggestions = []
+        self.selected_suggestion_index = 0
+        self.suggestion_colors = {}
+
+
+    def handle_suggestion_event(self, event):
+        keysym = event.keysym
+        if keysym == "Tab":
+            if self.selected_suggestion_index < len(self.suggestions):
+                selected_suggestion = self.suggestions[self.selected_suggestion_index]
+                if isinstance(selected_suggestion, tuple):
+                    selected_suggestion = selected_suggestion[0]
+                self.insert_selected_suggestion(selected_suggestion.strip())
+            self.clear_suggestions()
+        elif keysym in ("Alt_L", "Alt_R"):
+            if self.suggestions and not self.is_alt_arrow_pressed:
+                self.selected_suggestion_index = (self.selected_suggestion_index - 1) % len(self.suggestions) if keysym == "Alt_R" else (self.selected_suggestion_index + 1) % len(self.suggestions)
+                self.highlight_suggestions()
+            self.is_alt_arrow_pressed = False
+        elif keysym in ("Up", "Down", "Left", "Right") or event.char == ",":
+            self.clear_suggestions()
+        else:
+            return False
+        return True
+
+
+    def update_suggestions(self, event=None):
+        if event is None:
+            event = type('', (), {})()
+            event.keysym = ''
+            event.char = ''
+        cursor_position = self.parent.text_box.index("insert")
+        if self.cursor_inside_tag(cursor_position):
+            self.clear_suggestions()
+            return
+        if self.handle_suggestion_event(event):
+            return
+        text = self.parent.text_box.get("1.0", "insert")
+        self.clear_suggestions()
+        if self.parent.last_word_match_var.get():
+            words = text.split()
+            current_word = words[-1] if words else ''
+        else:
+            if self.parent.list_mode_var.get():
+                elements = [element.strip() for element in text.split('\n')]
+            else:
+                elements = [element.strip() for element in text.split(',')]
+            current_word = elements[-1]
+        current_word = current_word.strip()
+        if current_word and (len(self.selected_csv_files) >= 1 or self.parent.use_mytags_var.get()):
+            suggestions = self.autocomplete.get_suggestion(current_word)
+            suggestions.sort(key=lambda x: self.autocomplete.get_score(x[0], current_word), reverse=True)
+            self.suggestions = [(suggestion[0].replace("_", " ") if suggestion[0] not in self.autocomplete.tags_with_underscore else suggestion[0], suggestion[1]) for suggestion in suggestions]
+            if self.suggestions:
+                self.highlight_suggestions()
+            else:
+                self.clear_suggestions()
+        else:
+            self.clear_suggestions()
+
+
+    def highlight_suggestions(self):
+        def on_mouse_hover(tag_name, highlight, event=None):
+            if highlight:
+                widget.tag_config(tag_name, relief='raised', borderwidth=1)
+                widget.config(cursor="hand2")
+            else:
+                widget.tag_config(tag_name, relief='flat', borderwidth=0)
+                widget.config(cursor="")
+        widget = self.parent.suggestion_textbox
+        widget.config(state='normal')
+        widget.delete('1.0', 'end')
+        configured_colors = set()
+        num_suggestions = len(self.suggestions)
+        for index, (suggestion_text, classifier_id) in enumerate(self.suggestions):
+            classifier_id = classifier_id[0]
+            color_index = int(classifier_id) % len(self.suggestion_colors) if classifier_id and classifier_id.isdigit() else 0
+            suggestion_color = self.suggestion_colors[color_index]
+            bullet_symbol = "⚫" if index == self.selected_suggestion_index else "⚪"
+            suggestion = f"suggestion_tag_{index}"
+            widget.insert('end', bullet_symbol)
+            widget.insert('end', f" {suggestion_text} ", (suggestion, suggestion_color))
+            if suggestion_color not in configured_colors:
+                widget.tag_config(suggestion_color, foreground=suggestion_color, font=('Segoe UI', '9'))
+                configured_colors.add(suggestion_color)
+            widget.tag_bind(suggestion, '<Button-1>', partial(self.on_suggestion_click, index))
+            widget.tag_bind(suggestion, '<Enter>', partial(on_mouse_hover, suggestion, True))
+            widget.tag_bind(suggestion, '<Leave>', partial(on_mouse_hover, suggestion, False))
+            if index < num_suggestions - 1:
+                widget.insert('end', ', ')
+        widget.config(state='disabled')
+
+
+    def cursor_inside_tag(self, cursor_position):
+        line, column = map(int, cursor_position.split('.'))
+        line_text = self.parent.text_box.get(f"{line}.0", f"{line}.end")
+        if self.parent.list_mode_var.get():
+            inside_tag = column not in (0, len(line_text))
+        else:
+            inside_tag = not (column == 0 or line_text[column-1:column] in (',', ' ') or line_text[column:column+1] in (',', ' ') or column == len(line_text))
+        return inside_tag
+
+
+    def clear_suggestions(self):
+        self.suggestions = []
+        self.selected_suggestion_index = 0
+        self.parent.suggestion_textbox.config(state='normal')
+        self.parent.suggestion_textbox.delete('1.0', 'end')
+        self.parent.suggestion_textbox.insert('1.0', "...")
+        self.parent.suggestion_textbox.config(state='disabled')
+
+
+# --------------------------------------
+# Insert Suggestion
+# --------------------------------------
+    def insert_selected_suggestion(self, selected_suggestion):
+        selected_suggestion = selected_suggestion.strip()
+        text = self.parent.text_box.get("1.0", "insert").rstrip()
+        if self.parent.last_word_match_var.get():
+            words = text.split()
+            current_word = words[-1] if words else ''
+        else:
+            elements = [element.strip() for element in text.split('\n' if self.parent.list_mode_var.get() else ',')]
+            current_word = elements[-1]
+        remaining_text = self.parent.text_box.get("insert", "end").rstrip('\n')
+        start_of_current_word = "1.0 + {} chars".format(len(text) - len(current_word))
+        self.parent.text_box.delete(start_of_current_word, "insert")
+        if not remaining_text.startswith(('\n' if self.parent.list_mode_var.get() else ',')) and not self.parent.last_word_match_var.get():
+            self.parent.text_box.insert(start_of_current_word, selected_suggestion + ('\n' if self.parent.list_mode_var.get() else ','))
+        else:
+            self.parent.text_box.insert(start_of_current_word, selected_suggestion)
+        if self.parent.list_mode_var.get() and remaining_text:
+            self.insert_newline_listmode(called_from_insert=True)
+            self.parent.text_box.mark_set("insert", "insert + 1 lines")
+
+
+    def insert_newline_listmode(self, event=None, called_from_insert=False):
+        if self.parent.list_mode_var.get():
+            self.parent.text_box.insert("insert", '\n')
+            if called_from_insert and self.parent.text_box.index("insert") != self.parent.text_box.index("end-1c"):
+                self.parent.text_box.mark_set("insert", "insert-1l")
+            return 'break'
+
+
+    def on_suggestion_click(self, suggestion_index, event=None):
+        selected_suggestion, _ = self.suggestions[suggestion_index]
+        self.insert_selected_suggestion(selected_suggestion.strip())
+        self.clear_suggestions()
+
+
+# --------------------------------------
+# Suggestion Settings
+# --------------------------------------
+    def update_autocomplete_dictionary(self):
+        csv_vars = {
+            'danbooru.csv': self.parent.csv_danbooru,
+            'danbooru_safe.csv': self.parent.csv_danbooru_safe,
+            'e621.csv': self.parent.csv_e621,
+            'dictionary.csv': self.parent.csv_english_dictionary,
+            'derpibooru.csv': self.parent.csv_derpibooru
+            }
+        self.selected_csv_files = [csv_file for csv_file, var in csv_vars.items() if var.get()]
+        if not self.selected_csv_files:
+            self.autocomplete = Autocomplete("None", include_my_tags=self.parent.use_mytags_var.get())
+        else:
+            self.autocomplete = Autocomplete(self.selected_csv_files[0], include_my_tags=self.parent.use_mytags_var.get())
+            for csv_file in self.selected_csv_files[1:]:
+                self.autocomplete.autocomplete_dict.update(Autocomplete(csv_file).autocomplete_dict, include_my_tags=self.parent.use_mytags_var.get())
+        self.clear_suggestions()
+        self.set_suggestion_color(self.selected_csv_files[0] if self.selected_csv_files else "None")
+        self.set_suggestion_threshold()
+
+
+    def set_suggestion_color(self, csv_file):
+        color_mappings = {
+            'None':                 {0: "black"},
+            'dictionary.csv':       {0: "black",    1: "black",     2: "black",     3: "black",     4: "black",     5: "black",     6: "black",     7: "black",     8: "black"},
+            'danbooru.csv':         {0: "black",    1: "#c00004",   2: "black",     3: "#a800aa",   4: "#00ab2c",   5: "#fd9200"},
+            'danbooru_safe.csv':    {0: "black",    1: "#c00004",   2: "black",     3: "#a800aa",   4: "#00ab2c",   5: "#fd9200"},
+            'e621.csv':             {-1: "black",   0: "black",     1: "#f2ac08",   3: "#dd00dd",   4: "#00aa00",   5: "#ed5d1f",   6: "#ff3d3d",   7: "#ff3d3d",   8: "#228822"},
+            'derpibooru.csv':       {0: "black",    1: "#e5b021",   3: "#fd9961",   4: "#cf5bbe",   5: "#3c8ad9",   6: "#a6a6a6",   7: "#47abc1",   8: "#7871d0",   9: "#df3647",   10: "#c98f2b",  11: "#e87ebe"}
+        }
+        black_mappings = {key: "black" for key in color_mappings[csv_file].keys()}
+        self.suggestion_colors = color_mappings[csv_file] if self.parent.colored_suggestion_var.get() else black_mappings
+
+
+    def set_suggestion_quantity(self, suggestion_quantity):
+        self.autocomplete.max_suggestions = suggestion_quantity
+        self.update_suggestions(event=None)
+
+
+    def set_suggestion_threshold(self):
+        thresholds = {
+            "Slow"  : 275000,
+            "Normal": 130000,
+            "Fast"  : 75000,
+            "Faster": 40000
+        }
+        self.autocomplete.suggestion_threshold = thresholds.get(self.parent.suggestion_threshold_var.get())
+
+
+    def clear_dictionary_csv_selection(self):
+        for attr in ['csv_danbooru', 'csv_derpibooru', 'csv_e621', 'csv_english_dictionary']:
+            getattr(self, attr).set(False)
+        self.update_autocomplete_dictionary()
